Remove call to sqlite3_initialize
The sqlite3_initialize() routine is called internally by many
other SQLite interfaces so that an application usually does not
need to invoke sqlite3_initialize() directly. For example,
sqlite3_open() calls sqlite3_initialize() so the SQLite library
will be automatically initialized when sqlite3_open() is called
if it has not be initialized already.

However, if SQLite is compiled with the SQLITE_OMIT_AUTOINIT
compile-time option, then the automatic calls to
sqlite3_initialize() are omitted and the application must call
sqlite3_initialize() directly prior to using any other SQLite
interface. For maximum portability, it is recommended that
applications always invoke sqlite3_initialize() directly prior
to using any other SQLite interface. Future releases of SQLite
may require this. In other words, the behavior exhibited when
SQLite is compiled with SQLITE_OMIT_AUTOINIT might become the
default behavior in some future release of SQLite.